chore: factor the filecard code
